UADS-19 fix code in wrong section
